Corrige achados de auditoria de problemas comuns/incomuns de NavMesh
Bateria de correcoes sobre o pacote com.mariochi.customnavmesh, cobrindo
os itens levantados em duas rodadas de auditoria (problemas classicos de
NavMesh e uma revisao critica do codigo novo desta sessao):

Avoidance / ORCA
- Substitui a soma de empurroes + media por ORCA de verdade (LinearProgram1/2/3),
  com determinismo garantido (ordena vizinhos por indice antes de montar as
  retas, independente da ordem de insercao paralela no spatial hash).
- Raio de deteccao escala automaticamente com o Radius do proprio agente
  (correcao de "parede invisivel" para agentes com raio bem maior que o
  Neighbor Query Radius configurado), e a janela de busca no grid acompanha.
- Radius negativo agora e clampado em todos os pontos de entrada
  (CustomNavMeshAgent/Obstacle, NavMeshJobManager, e defensivamente no Job).

Obstaculos dinamicos
- Novo CustomNavMeshObstacle: avoidance nao-reciproca contra objetos que nao
  sao agentes (fisica, animacao, scripts de terceiros), no mesmo spatial hash
  dos agentes via indice negativo.

NavMeshLink (off-mesh links)
- NavMeshGraphBuilder.BuildLinks coleta NavMeshLink da cena e monta arestas
  virtuais extras no A*; corredor final insere saltos retos entre triangulos
  nao-adjacentes, com bypass do clamp de superficie durante a travessia.
- NavMeshLink.width agora e usado (deslocamento lateral deterministico por
  agente, validado contra o grid espacial).
- StalenessWatcher (deteccao automatica de grafo desatualizado) agora tambem
  cobre mudancas de NavMeshLink, nao so da triangulacao.

Pathfinding / corredor
- Corredor de "melhor esforco" quando o destino e inalcancavel (ilha
  desconectada), em vez de deixar o agente parado.
- Funil aceita raio do agente e encolhe portais internos pra manter distancia
  minima de paredes/vertices (reduz corner-clipping visual).
- ClampToNavMesh ganha uma camada de BFS por adjacencia real entre o cache
  local e o fallback irrestrito de grid, evitando que o clamp "atravesse"
  uma parede fina em situacoes de empurrao/velocidade alta.

Robustez numerica e de memoria
- Chave da hash espacial (grid de triangulos e de agentes/obstaculos) trocada
  de int (overflow silencioso em mundos grandes) para long (empacotamento
  exato, sem colisao possivel).
- Triangulo degenerado tratado explicitamente em ClosestPointOnTriangle;
  epsilon adicionado as comparacoes de area no funil.
- WaypointReachDistance clampado a um piso positivo (evitava agente preso
  orbitando o mesmo waypoint pra sempre).
- Grid de triangulos calibrado automaticamente (media de aresta ponderada
  por area, nao por contagem de triangulo).

Ciclo de vida / multi-instancia
- Transform movido por fora da API (fisica/animacao) e adotado como nova
  posicao valida em vez de silenciosamente sobrescrito.
- NativeArrays persistentes protegidos contra vazamento entre sessoes de
  Play sem Domain Reload (guarda em Awake + hook de Editor).
- Segunda instancia de NavMeshJobManager nao e mais destruida — sobrevive e
  funciona de forma independente; campo Manager em
  CustomNavMeshAgent/Obstacle permite apontar pra uma instancia especifica
  (multi-cena aditiva).

Build
- NavMeshLink e Unity.AI.Navigation.NavMeshLink (pacote com.unity.ai.navigation),
  nao UnityEngine.AI — using, referencia de asmdef e dependencia de
  package.json corrigidos.

READM.md atualizado com todas as secoes/limitacoes correspondentes.

+++ Packages/com.mariochi.customnavmesh/Runtime/Jobs/BuildAgentSpatialHashJob.cs
@@ -16,12 +16,12 @@ namespace CustomNavMesh
     {
         [ReadOnly] public NativeArray<float3> Positions;
         public float CellSize;
-        public NativeParallelMultiHashMap<int, int>.ParallelWriter HashWriter;
+        public NativeParallelMultiHashMap<long, int>.ParallelWriter HashWriter;
 
         public void Execute(int index)
         {
             int2 cell = (int2)math.floor(Positions[index].xz / CellSize);
-            int key = AvoidanceAndMoveJob.HashCellKey(cell);
+            long key = AvoidanceAndMoveJob.HashCellKey(cell);
             HashWriter.Add(key, index);
         }
     }
